Refactoring: chess-independent utils extracted
